Give menu contents a higher z-index
Minor thing, but if you follow the approach of using a negative `top` value to overlay something on the image, it ends up being on top of the dropdown menu. To avoid this, we can move the menu to a z-index of 3 - there doesn't seem to be any case where we'd want something to cover up the menu anyway, so the approach seems sound.