DE30421 Fix if condition
Found while fixing DE30421, although not strictly related to it. The typo in the `if` condition here meant that even if `href` was null or undefined, the `d2l-image-tile` would always render the `<a>`. This caused weird tabbing behaviour, as the focus would move to an anchor without an href, which doesn't really make a whole lot of sense.